+++ src/tekhsi/helpers/__init__.py
@@ -1,9 +1,10 @@
 """Helpers used by the `TekHSI` package."""
 
 from tekhsi.helpers.constants import PACKAGE_NAME
-from tekhsi.helpers.functions import print_with_timestamp
+from tekhsi.helpers.logging import configure_logging, LoggingLevels
 
 __all__ = [
+    "configure_logging",
+    "LoggingLevels",
     "PACKAGE_NAME",
-    "print_with_timestamp",
 ]